move flow components into its own subfolder, separate LoginGate from Login
