+++ src/components/App.tsx
@@ -1,13 +1,13 @@
 import { Text } from "ink"
 import React from "react"
 
+import { List } from "../flows/List.js"
+import { Logout } from "../flows/Logout.js"
+import { Start } from "../flows/Start/Start.js"
 import { cliCommand, cli } from "../lib/cli.js"
 import { packageJson } from "../lib/packageJson.js"
 
-import { List } from "./List.js"
 import { LoginGate } from "./LoginGate.js"
-import { Logout } from "./Logout.js"
-import { Start } from "./Start/Start.js"
 
 export function App() {
   const { command, commandInvalid } = cliCommand()
